pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "edupptx"
version = "0.1.0"
description = "AI-powered educational presentation generator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cairosvg>=2.9.0",
    "click>=8.3.2",
    "openai>=2.31.0",
    "pillow>=12.2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-pptx>=1.0.2",
]

[project.scripts]
edupptx = "edupptx.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
